requirements.txt
bcrypt==3.2.0
cffi==2.0.0
inquirerpy==0.3.4
mariadb==1.1.14
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==25.0
passlib==1.7.4
pfzy==0.3.4
prompt_toolkit==3.0.52
pycparser==2.23
Pygments==2.19.2
python-dotenv==1.1.1
rich==14.2.0
six==1.17.0
wcwidth==0.2.14
